requirements.txt
fastapi==0.112.1
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
numpy==1.26.4
pillow==10.3.0
opencv-python-headless==4.10.0.84
pydantic-settings==2.4.0